libfwxconfig: workaround for multihead breakage
the problem is the following: when we have two vga outputs and the
second has limited capabilities (no hw accel, for example), then the
first should use a real driver, the second can be just "vga".

now X -configure sadly sets the wrong order by default in the generated
xorg.conf: the first is vga and the second is the real driver, so we
have to restore the right order in fwx_doconfig().